SWANDROID-3325: [PM] STB reboots during shutdown sequence
The BT_HOST_WAKE signal from the BT chip was inadvertently waking
the STB up during shutdown.  The simple fix is to ensure that the
BT_DEV_WAKE signal is left HIGH when entering S5.  This now allows
the STB to fully shutdown using either:

a) Long-press of the POWER button method.
b) Setting the Android property "persist.sys.power.offstate" to "S5
   and doing a normal (short) press of the POWER button.

refs #SWANDROID-3325

Change-Id: I4fa2b9a4caa8f0bc0af8f8d7ff4d3a0d4e7a1569