avko: increase the lmk threshold.
by increasing the lmk threshold, we want to force memory reclaim
to kick in earlier which helps with scenarios where a lots of applications
are open one after the other continuously leading to device becoming
sluggish as reclaim/swap takes place.  by increasing the limit of the lmk
we allow more background killing to improve the reclaim/swap time and
need; overall giving a better user experience.

b/29521215

Change-Id: Id35d23238f00c00230b545d24a8cfc1227fe3a6d
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>